# [FIX] stock: improve SQL view perf by inlining CTE.