Report stock quantity defines a custom SQL View in init function.
Inlining the view CTE improves performances of the
product.template/product.product forecasted quantity reports.

For postgresql >= 12.0, the inlining of CTEs is automatic. However, since
odoo.sh is currently on psql 10.15 , there are still performance improvements
that can be attained by manually inlining CTEs in custom views.

When generating the report from the product_template form view, the domain returned
by `_getReportDomain` of `report_stock_forecasted.js` contains `[('product_tmpl_id', '=', this.productId)]`.
Since `report.stock.quantity.product_tmpl_id` is `related` to `report.stock.quantity.product_id`, 
the generated `WHERE` clause contains a subquery to retrieve the product_ids for the given
product_tmpl_id.

To avoid that, this PR removes the related attribute of product_tmpl_id and changes the SQL View by 
adding product_tmpl_id to the selected/grouped_by columns (see ryv comment below).

##### Speedup

In a client database with 200K stock moves, 10K stock quant, 5K product templates and 13K products.

###### Only with CTE inlining

- Speedup for forecasted report button in product_template form view: 21s -> 16s.
- Speedup for forecasted report button in product_product form view: 7.50s -> 60ms.

###### With CTE inlining + Product Template in View

- Speedup for forecasted report button in product_template form view: 21s -> 65ms.
- Speedup for forecasted report button in product_product form view: 7.50s -> 65ms.

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr