@yelizariev 2021-04-08
Did you measure speed up for each update separately?