@ryv-odoo 2021-04-16
Hello @Aurelienvd , Thank you for the PR.
I am totally agree with the CTE inlining (nice improvements).

For the JS RPC to get the product_ids, it is a little bit "pity" to add a rpc request to improve the performance of the report. Maybe a other solution to avoid this extra request: remove the related of `product_tmpl_id` field, and change the SQL request (add the product_tmpl_id in the select + group_by). What do you think ? It will avoid the subquery and it shouldn't impact a lot the performance of SQL view (to test).