@Aurelienvd 2021-04-16
> 
> 
> Hello @Aurelienvd , Thank you for the PR.
> I am totally agree with the CTE inlining (nice improvements).
> 
> For the JS RPC to get the product_ids, it is a little bit "pity" to add a rpc request to improve the performance of the report. Maybe a other solution to avoid this extra request: remove the related of `product_tmpl_id` field, and change the SQL request (add the product_tmpl_id in the select + group_by). What do you think ? It will avoid the subquery and it shouldn't impact a lot the performance of SQL view (to test).

Yeah I agree. I didn't test it yet but as you, and Ivan, pointed out, adding the _rpc does induce some uncertainty in the performances gain if the customer connection is janky or if there are lots of variants to send to the client. Changing the SQL request does look much better. I'll see if I can manage to make it work.