@Aurelienvd 2021-04-20
Hello @ryv-odoo , I removed the _rpc call and changed the SQL View as suggested, I think it's even faster for the product_template 😮 now.