@ryv-odoo 2021-04-20
Very nice improvements. Thank you for your work @Aurelienvd , it is always appreciated!

@robodoo  r+